requirements.txt
customtkinter
pandas
openpyxl
pdfkit
Pillow
